requirements.txt
streamlit
face_recognition
opencv-python
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
Pillow